requirements.txt
PyYAML==6.0.2
paramiko==3.5.1
fastapi==0.115.12
uvicorn==0.34.0
jinja2==3.1.6
python-multipart==0.0.20
itsdangerous==2.2.0
